refactor: move timestamp parsing into runtime readmodel
